pyproject.toml
[project]
name = "b2b-saas-graphrag"
version = "0.1.0"
description = "Graph-enhanced RAG for B2B SaaS product analytics with synthetic data generation"
requires-python = ">=3.10"

dependencies = [
    "networkx>=3.2",
    "pandas>=2.1",
    "numpy>=1.26",
    "faker>=22.0",
    "sentence-transformers>=2.2",
    "faiss-cpu>=1.7",
    "anthropic>=0.40",
    "python-dotenv>=1.0",
    "tqdm>=4.66",
    "datasette>=0.65",
    "pyvis>=0.3",
]

[project.optional-dependencies]
dev = ["pytest>=7.4", "black>=24.1", "ruff>=0.1"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
